Relocate QR code to  named module
And clean up a redundant conditional
